abstract the delete functionality of project detail fields and add the confirmation box of deletion
