+++ package.json
@@ -15,6 +15,7 @@
 		"html-webpack-plugin": "^2.28.0",
 		"lodash.times": "^4.3.2",
 		"react": "^16.12.0",
+		"react-confirm-alert": "^2.6.1",
 		"react-datepicker": "^2.12.1",
 		"react-dom": "^16.12.0",
 		"react-helmet": "^5.2.1",
